Revert "Swiping on the top of full screen apps shows status bar for 10 seconds"
This reverts commit e13d91bd9fdc4e1bc95aa382758a1fee55736b00.

I reverted this as it really doesn't play nice with AutoHide, I thought it would compliment it in some sort of way but I was so wrong. I would really like someway of activating the statusbar with a swipe when AutoHide is on as at the moment you cannot access your tiles unless you go from the lockscreen. Slim has an Enhanced Toggle Notification Feature but this only allows the statusbar to drop if you
assign a button to toggle the notification. I don't want that sort of work around.